Remove defailt AllCops/Exclude as these now clash with project rubocop.yml
In Rubocop 0.53.0 specifying AllCops/Exclude in here (pulled in via default.yml in the project) causes an error if is also specified in .rubocop.yml in the project dir. Really rubocop should merge them. Hey ho So for now remove the default Excludes here and add to the projects rubocop.yml.